+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 This document details the changes between each release.
 
+## [1.1.0]
+
+### Added
+
+* OSC bundle support, via a simple `OSCBundle` container class.
+
+### Changed
+
+### Removed
+
+* Default copy constructors are removed.
+
+### Fixed
+
+* The constructors now check if the initial memory allocation failed.
+
 ## LiteOSCParser v1.0.0
 
 This is the first LiteOSCParser release. It provides a way to fix the amount